chore: split bind out from serve, make eventual gateway generation into a function
