+++ cli/crates/production-server/src/server.rs
@@ -2,27 +2,25 @@ mod cors;
 mod csrf;
 mod engine;
 mod gateway;
+mod graph_fetch_method;
 mod graph_updater;
 mod state;
 
-use crate::{config::Config, GraphFetchMethod};
+pub use graph_fetch_method::GraphFetchMethod;
+
+use crate::config::{Config, TlsConfig};
 use axum::{routing::get, Router};
 use axum_server::tls_rustls::RustlsConfig;
 use gateway_v2::local_server::{WebsocketAccepter, WebsocketService};
 use state::ServerState;
-use std::{
-    net::{IpAddr, Ipv4Addr, SocketAddr},
-    sync::Arc,
-};
-use tokio::sync::{mpsc, watch};
+use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+use tokio::sync::mpsc;
 use tower_http::{cors::CorsLayer, services::ServeDir};
-
-use self::graph_updater::GraphUpdater;
 
 pub(super) async fn serve(
     listen_addr: Option<SocketAddr>,
     config: Config,
-    graph: GraphFetchMethod,
+    fetch_method: GraphFetchMethod,
 ) -> crate::Result<()> {
     let path = config.graph.path.as_deref().unwrap_or("/graphql");
 
@@ -30,42 +28,11 @@ pub(super) async fn serve(
         .or(config.network.listen_address)
         .unwrap_or(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 4000));
 
-    let (sender, gateway) = watch::channel(None);
-
-    match graph {
-        GraphFetchMethod::FromApi {
-            access_token,
-            graph_name,
-            branch,
-        } => {
-            tokio::spawn(async move {
-                let mut updater = GraphUpdater::new(&graph_name, branch.as_deref(), access_token, sender)?
-                    .enable_introspection(config.graph.introspection);
-
-                if let Some(operation_limits) = config.operation_limits {
-                    updater = updater.with_operation_limits(operation_limits);
-                }
-
-                if let Some(auth_config) = config.authentication {
-                    updater = updater.with_authentication(auth_config);
-                }
-
-                updater.poll().await;
-
-                Ok::<_, crate::Error>(())
-            });
-        }
-        GraphFetchMethod::FromLocal { federated_schema } => {
-            let gateway = gateway::generate(
-                &federated_schema,
-                config.operation_limits,
-                config.authentication,
-                config.graph.introspection,
-            )?;
-
-            sender.send(Some(Arc::new(gateway)))?;
-        }
-    }
+    let gateway = fetch_method.into_gateway(
+        config.graph.introspection,
+        config.operation_limits,
+        config.authentication,
+    )?;
 
     let (websocket_sender, websocket_receiver) = mpsc::channel(16);
     let websocket_accepter = WebsocketAccepter::new(websocket_receiver, gateway.clone());
@@ -91,9 +58,15 @@ pub(super) async fn serve(
         router = csrf::inject_layer(router);
     }
 
+    bind(addr, &path, router, config.tls).await?;
+
+    Ok(())
+}
+
+async fn bind(addr: SocketAddr, path: &str, router: Router, tls: Option<TlsConfig>) -> Result<(), crate::Error> {
     let app = router.into_make_service();
 
-    match config.tls {
+    match tls {
         Some(ref tls) => {
             tracing::info!("starting the Grafbase gateway in https://{addr}{path}");
